example_python_project_3/requirements.txt
requests>=2.26.0 
googlesearch_python==1.1.0